# server/openapi.yaml
openapi: 3.0.3
info:
  title: UCC-MCA Intelligence API
  description: |
    AI-powered lead generation system for Merchant Cash Advance providers.
    Analyzes Uniform Commercial Code (UCC) filings to identify businesses
    with active financing and predict MCA likelihood.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: Proprietary
    url: https://example.com/license
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.example.com
    description: Production server
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /api/prospects:
    get:
      tags:
        - Prospects
      summary: List prospects
      description: Returns paginated, filtered, and sorted list of UCC prospects
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: state
          in: query
          description: Filter by US state code (2 characters)
          schema:
            type: string
            minLength: 2
            maxLength: 2
            example: CA
        - name: industry
          in: query
          description: Filter by industry
          schema:
            $ref: '#/components/schemas/Industry'
        - name: min_score
          in: query
          description: Minimum priority score filter
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - name: max_score
          in: query
          description: Maximum priority score filter
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - name: status
          in: query
          description: Filter by prospect status
          schema:
            type: string
            enum: [all, unclaimed, claimed, contacted]
        - name: sort_by
          in: query
          description: Field to sort by
          schema:
            type: string
            enum: [priority_score, created_at, company_name]
            default: priority_score
        - name: sort_order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of prospects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProspectListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'

    post:
      tags:
        - Prospects
      summary: Create prospect
      description: Create a new UCC prospect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProspectRequest'
      responses:
        '201':
          description: Prospect created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prospect'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/competitors:
    get:
      tags:
        - Competitors
      summary: List competitors
      description: Returns list of competitor secured parties
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: state
          in: query
          schema:
            type: string
            minLength: 2
            maxLength: 2
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [filing_count, total_amount, name]
            default: filing_count
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of competitors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompetitorListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/portfolio:
    get:
      tags:
        - Portfolio
      summary: List portfolio companies
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: health_grade
          in: query
          schema:
            type: string
            enum: [A, B, C, D, F]
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [funded_date, health_score, company_name]
            default: funded_date
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of portfolio companies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/jobs/enrichment:
    post:
      tags:
        - Jobs
      summary: Trigger enrichment job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - prospectIds
              properties:
                prospectIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                  minItems: 1
                  maxItems: 100
                force:
                  type: boolean
                  default: false
      responses:
        '201':
          description: Job queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/contacts:
    get:
      tags:
        - Contacts
      summary: List contacts
      parameters:
        - name: org_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: search
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            $ref: '#/components/schemas/ContactRole'
        - name: tags
          in: query
          schema:
            type: string
            description: Comma-separated list of tags
        - name: is_active
          in: query
          schema:
            type: boolean
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [first_name, last_name, created_at, last_contacted_at]
            default: last_name
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: List of contacts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Contacts
      summary: Create contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContactRequest'
      responses:
        '201':
          description: Contact created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/deals:
    get:
      tags:
        - Deals
      summary: List deals
      parameters:
        - name: org_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: stage_id
          in: query
          schema:
            type: string
            format: uuid
        - name: assigned_to
          in: query
          schema:
            type: string
            format: uuid
        - name: prospect_id
          in: query
          schema:
            type: string
            format: uuid
        - name: priority
          in: query
          schema:
            $ref: '#/components/schemas/DealPriority'
        - name: search
          in: query
          schema:
            type: string
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created_at, updated_at, amount_requested, expected_close_date]
            default: created_at
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of deals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Deals
      summary: Create deal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDealRequest'
      responses:
        '201':
          description: Deal created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deal'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Issued API key as a JWT bearer token

    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Issued API key header alternative to bearer auth

    apiKeyIssuerSecret:
      type: apiKey
      in: header
      name: X-API-Key-Issuer-Secret
      description: Operator-only secret used to issue customer API keys

    twilioSignature:
      type: apiKey
      in: header
      name: X-Twilio-Signature
      description: Twilio request signature

    sendgridSignature:
      type: apiKey
      in: header
      name: X-Twilio-Email-Event-Webhook-Signature
      description: SendGrid webhook signature

    plaidSignature:
      type: apiKey
      in: header
      name: Plaid-Verification
      description: Plaid webhook signature
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - message
            - code
          properties:
            message:
              type: string
            code:
              type: string
            statusCode:
              type: integer
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    Industry:
      type: string
      enum:
        - restaurant
        - retail
        - construction
        - healthcare
        - manufacturing
        - services
        - technology
    ContactRole:
      type: string
      enum:
        - owner
        - ceo
        - cfo
        - controller
        - manager
        - bookkeeper
        - other
...
    ContactMethod:
      type: string
      enum:
        - email
        - phone
        - mobile
        - sms
...
    DealPriority:
      type: string
      enum:
        - low
        - normal
        - high
        - urgent
...
    UseOfFunds:
      type: string
      enum:
        - working_capital
        - inventory
        - equipment
        - expansion
        - payroll
        - marketing
        - debt_consolidation
        - real_estate
        - other

    # ===========================================
    # Health Schemas
    # ===========================================
...
    Prospect:
      type: object
      properties:
        id:
          type: string
          format: uuid
        company_name:
          type: string
        state:
          type: string
        industry:
          $ref: '#/components/schemas/Industry'
        priority_score:
          type: integer
          minimum: 0
          maximum: 100
        lien_amount:
          type: number
        filing_date:
          type: string
          format: date-time
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProspectListResponse:
      type: object
      properties:
        prospects:
          type: array
          items:
            $ref: '#/components/schemas/Prospect'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    CreateProspectRequest:
      type: object
      required:
        - company_name
        - state
        - industry
      properties:
        company_name:
          type: string
          minLength: 1
        state:
          type: string
          minLength: 2
          maxLength: 2
        industry:
          $ref: '#/components/schemas/Industry'
        lien_amount:
          type: number
        filing_date:
          type: string
          format: date-time
...
    Competitor:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        filing_count:
          type: integer
        total_amount:
          type: number
        states:
          type: array
          items:
            type: string
    CompetitorListResponse:
      type: object
      properties:
        competitors:
          type: array
          items:
            $ref: '#/components/schemas/Competitor'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    PortfolioCompany:
      type: object
      properties:
        id:
          type: string
          format: uuid
        company_name:
          type: string
        funded_date:
          type: string
          format: date
        funded_amount:
          type: number
        health_score:
          type: integer
        health_grade:
          type: string
          enum: [A, B, C, D, F]
    PortfolioListResponse:
      type: object
      properties:
        companies:
          type: array
          items:
            $ref: '#/components/schemas/PortfolioCompany'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    JobResponse:
      type: object
      properties:
        jobId:
          type: string
        queueName:
          type: string
        data:
          type: object
        status:
          type: string
...
    Contact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        mobile:
          type: string
        title:
          type: string
        role:
          $ref: '#/components/schemas/ContactRole'
        preferred_contact_method:
          $ref: '#/components/schemas/ContactMethod'
        timezone:
          type: string
        tags:
          type: array
          items:
            type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        last_contacted_at:
          type: string
          format: date-time
...
    ContactListResponse:
      type: object
      properties:
        contacts:
          type: array
          items:
            $ref: '#/components/schemas/Contact'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CreateContactRequest:
      type: object
      required:
        - org_id
        - first_name
        - last_name
      properties:
        org_id:
          type: string
          format: uuid
        first_name:
          type: string
          minLength: 1
        last_name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: string
        phone_ext:
          type: string
        mobile:
          type: string
        title:
          type: string
        role:
          $ref: '#/components/schemas/ContactRole'
        preferred_contact_method:
          $ref: '#/components/schemas/ContactMethod'
        timezone:
          type: string
          default: America/New_York
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
        source:
          type: string
        created_by:
          type: string
          format: uuid
...
    Deal:
      type: object
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        prospect_id:
          type: string
          format: uuid
        contact_id:
          type: string
          format: uuid
        stage_id:
          type: string
          format: uuid
        assigned_to:
          type: string
          format: uuid
        amount_requested:
          type: number
        amount_approved:
          type: number
        term_months:
          type: integer
        factor_rate:
          type: number
        daily_payment:
          type: number
        weekly_payment:
          type: number
        use_of_funds:
          $ref: '#/components/schemas/UseOfFunds'
        priority:
          $ref: '#/components/schemas/DealPriority'
        probability:
          type: integer
          minimum: 0
          maximum: 100
        expected_close_date:
          type: string
          format: date
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    DealListResponse:
      type: object
      properties:
        deals:
          type: array
          items:
            $ref: '#/components/schemas/Deal'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CreateDealRequest:
      type: object
      required:
        - org_id
      properties:
        org_id:
          type: string
          format: uuid
        prospect_id:
          type: string
          format: uuid
        contact_id:
          type: string
          format: uuid
        stage_id:
          type: string
          format: uuid
        assigned_to:
          type: string
          format: uuid
        amount_requested:
          type: number
        term_months:
          type: integer
        use_of_funds:
          $ref: '#/components/schemas/UseOfFunds'
        use_of_funds_details:
          type: string
        priority:
          $ref: '#/components/schemas/DealPriority'
        expected_close_date:
          type: string
          format: date-time
...